requirements.txt
sounddevice==0.4.7
numpy==1.26.4
openai==1.30.5
python-dotenv==1.0.1
openai-whisper==20231117
torch==2.2.2